requirements.txt
aiosmtplib==3.0.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-mail==1.4.2
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
Jinja2==3.1.4
MarkupSafe==3.0.2
passlib==1.7.4
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pypika-tortoise==0.3.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.41.3
tortoise-orm==0.22.1
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
